+++ P2PChat/P2PChat/Migrations/ApplicationContextModelSnapshot.cs
@@ -21,7 +21,7 @@ namespace P2PChat.Migrations
 
             modelBuilder.Entity("P2PChat.Models.Message", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
